+++ src/main/java/com/codetaylor/mc/athenaeum/registry/strategy/IForgeRegistryEventRegistrationStrategy.java
@@ -1,0 +1,9 @@
+package com.codetaylor.mc.athenaeum.registry.strategy;
+
+import net.minecraftforge.registries.IForgeRegistry;
+import net.minecraftforge.registries.IForgeRegistryEntry;
+
+public interface IForgeRegistryEventRegistrationStrategy<T extends IForgeRegistryEntry<T>> {
+
+  void register(IForgeRegistry<T> forgeRegistry);
+}