+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 1.12.6
+* Added: biome registration strategy
 * Changed: combined helper package into util package
 * Changed: generalized client registration strategy
 * Changed: refactored block registry methods